add --color flag to lsstswBuild.sh
Disable ANSI color codes unless the --color flag is set.  The shell buildbot
runs lsstswBuild.sh under reports that it is a tty and the terminal type
supports colors but raw color codes appears in the stdout log.  This appears to
prevent reasonable automatic determination of terminal color code support.